refactor: sync the signaling list with pushnotification
